# Conflicts:
#	src/main/java/net/geant/nmaas/portal/domain/RejectionReason.java
#	src/main/java/net/geant/nmaas/portal/domain/ResourcesLimitValidationResult.java
#	src/main/java/net/geant/nmaas/portal/persistence/repositories/ResourcesLimitRepository.java
#	src/main/java/net/geant/nmaas/portal/service/ResourcesLimitService.java
#	src/main/java/net/geant/nmaas/portal/service/impl/ResourcesLimitServiceImpl.java
#	src/test/java/net/geant/nmaas/portal/service/impl/ResourcesLimitTest.java

+++ src/main/java/net/geant/nmaas/portal/service/impl/WebhookEventService.java
@@ -1,13 +1,16 @@
 package net.geant.nmaas.portal.service.impl;
 
-import net.geant.nmaas.portal.api.domain.WebhookEventDto;
+import lombok.RequiredArgsConstructor;
+import net.geant.nmaas.portal.domain.WebhookEventDto;
 import net.geant.nmaas.portal.api.exceptions.MissingElementException;
 import net.geant.nmaas.portal.api.security.EncryptionService;
-import net.geant.nmaas.portal.persistent.entity.WebhookEvent;
-import net.geant.nmaas.portal.persistent.entity.WebhookEventType;
-import net.geant.nmaas.portal.persistent.repositories.WebhookEventRepository;
+import net.geant.nmaas.portal.persistence.entity.Domain;
+import net.geant.nmaas.portal.persistence.entity.Role;
+import net.geant.nmaas.portal.persistence.entity.WebhookEvent;
+import net.geant.nmaas.portal.persistence.entity.WebhookEventType;
+import net.geant.nmaas.portal.persistence.repositories.WebhookEventRepository;
+import net.geant.nmaas.portal.service.UserService;
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.security.GeneralSecurityException;
@@ -15,19 +18,15 @@ import java.util.List;
 import java.util.stream.Stream;
 
 @Service
+@RequiredArgsConstructor
 public class WebhookEventService {
+
+    private static final String WEBHOOK_EVENT_NOT_FOUND = "WebhookEvent not found.";
 
     private final WebhookEventRepository webhookRepository;
     private final EncryptionService encryptionService;
     private final ModelMapper modelMapper;
-    private static final String WEBHOOK_EVENT_NOT_FOUND = "WebhookEvent not found.";
-
-    @Autowired
-    public WebhookEventService(WebhookEventRepository webhookRepository, EncryptionService encryptionService, ModelMapper modelMapper) {
-        this.webhookRepository = webhookRepository;
-        this.encryptionService = encryptionService;
-        this.modelMapper = modelMapper;
-    }
+    private final UserService userService;
 
     public WebhookEvent create(WebhookEventDto webhookEventDto) throws GeneralSecurityException {
         WebhookEvent webhookEvent = new WebhookEvent();
@@ -35,28 +34,43 @@ public class WebhookEventService {
         return webhookRepository.save(webhookEvent);
     }
 
-    public WebhookEventDto update(WebhookEventDto webhookEventDto) throws GeneralSecurityException {
+    public WebhookEventDto update(WebhookEventDto webhookEventDto, String username) throws GeneralSecurityException {
         WebhookEvent webhookEvent = webhookRepository.findById(webhookEventDto.getId())
                 .orElseThrow(() -> new MissingElementException(WEBHOOK_EVENT_NOT_FOUND));
-        setWebhookEvent(webhookEvent, webhookEventDto);
-        webhookEvent = webhookRepository.save(webhookEvent);
-        WebhookEventDto dto = modelMapper.map(webhookEvent, WebhookEventDto.class);
-        dto.setTokenValue(webhookEvent.getTokenValue() == null ? null : encryptionService.decrypt(webhookEvent.getTokenValue()));
-        return dto;
+      //  for update domain admin can not change domain of webhook event
+        if (userService.isAdmin(username) || (userService.isUserAdminInAnyDomain(webhookEvent.getDomain() != null ? List.of(webhookEvent.getDomain()) : List.of(), username)
+                && webhookEventDto.getDomain() != null && webhookEvent.getDomain().getId().equals(webhookEventDto.getDomain().getId()))) {
+            setWebhookEvent(webhookEvent, webhookEventDto);
+            webhookEvent = webhookRepository.save(webhookEvent);
+            WebhookEventDto dto = modelMapper.map(webhookEvent, WebhookEventDto.class);
+            dto.setTokenValue(webhookEvent.getTokenValue() == null ? null : encryptionService.decrypt(webhookEvent.getTokenValue()));
+            return dto;
+        } else {
+            throw new IllegalArgumentException("No access to webhook " + webhookEvent.getId());
+        }
     }
 
     private void setWebhookEvent(WebhookEvent webhookEvent, WebhookEventDto webhookEventDto) throws GeneralSecurityException {
+        //domain is combined only with APPLICATION_DEPLOYMENT, APPLICATION_REMOVAL and USER_ASSIGNMENT
+        if (webhookEventDto.getDomain() != null && Stream.of(WebhookEventType.DOMAIN_ACTION, WebhookEventType.DOMAIN_GROUP_ACTION).anyMatch(x -> webhookEventDto.getEventType().equals(x)))
+            throw new IllegalArgumentException("Domain can not combine with " + webhookEventDto.getEventType());
+
         webhookEvent.setName(webhookEventDto.getName());
         webhookEvent.setTargetUrl(webhookEventDto.getTargetUrl());
         webhookEvent.setEventType(webhookEventDto.getEventType());
         webhookEvent.setTokenValue(webhookEventDto.getTokenValue() == null ? null : encryptionService.encrypt(webhookEventDto.getTokenValue()));
         webhookEvent.setAuthorizationHeader(webhookEventDto.getAuthorizationHeader());
+        webhookEvent.setDomain(webhookEventDto.getDomain() != null ? new Domain(webhookEventDto.getDomain().getId()): null);
     }
 
-    public void remove(Long id) {
+    public void remove(Long id, String username) {
         WebhookEvent webhookEvent = webhookRepository.findById(id)
                 .orElseThrow(() -> new MissingElementException(WEBHOOK_EVENT_NOT_FOUND));
-        webhookRepository.delete(webhookEvent);
+        if (checkPrivileges(webhookEvent.getId(), username, webhookEvent.getDomain())) {
+            webhookRepository.delete(webhookEvent);
+        } else {
+            throw new IllegalArgumentException("No access to webhook " + webhookEvent.getId());
+        }
     }
 
     public List<WebhookEventDto> getAllWebhooks() {
@@ -82,5 +96,21 @@ public class WebhookEventService {
         WebhookEventDto dto = modelMapper.map(event, WebhookEventDto.class);
         dto.setTokenValue(event.getTokenValue() == null ? null : encryptionService.decrypt(event.getTokenValue()));
         return dto;
+    }
+
+    public WebhookEventDto getById(Long id, String username) throws GeneralSecurityException {
+        WebhookEvent event = webhookRepository.findById(id)
+                .orElseThrow(() -> new MissingElementException(String.format("WebhookEventType with id: %d cannot be found", id)));
+        if (checkPrivileges(id, username, event.getDomain())) {
+            WebhookEventDto dto = modelMapper.map(event, WebhookEventDto.class);
+            dto.setTokenValue(event.getTokenValue() == null ? null : encryptionService.decrypt(event.getTokenValue()));
+            return dto;
+        } else {
+            throw new IllegalArgumentException("No access to webhook " + id);
+        }
+    }
+
+    private boolean checkPrivileges(Long id, String username, Domain domain) {
+        return userService.isAdmin(username) || userService.isUserAdminInAnyDomain(domain != null ? List.of(domain) : List.of(), username);
     }
 }
